# [VL] Optimize Velox to Internal Row