--- a/backends-clickhouse/src/main/scala/io/glutenproject/execution/CHHashJoinExecTransformer.scala
+++ b/backends-clickhouse/src/main/scala/io/glutenproject/execution/CHHashJoinExecTransformer.scala
@@ -16,10 +16,14 @@
  */
 package io.glutenproject.execution
 
+import io.glutenproject.utils.CHJoinValidateUtil
+
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.optimizer.BuildSide
 import org.apache.spark.sql.catalyst.plans._
 import org.apache.spark.sql.execution.SparkPlan
+
+import io.substrait.proto.JoinRel
 
 case class CHShuffledHashJoinExecTransformer(
     leftKeys: Seq[Expression],
@@ -39,11 +43,21 @@ case class CHShuffledHashJoinExecTransformer(
     left,
     right,
     isSkewJoin) {
-
   override protected def withNewChildrenInternal(
       newLeft: SparkPlan,
       newRight: SparkPlan): CHShuffledHashJoinExecTransformer =
     copy(left = newLeft, right = newRight)
+
+  override def doValidateInternal(): Boolean = {
+    var shouldFallback = false
+    if (substraitJoinType != JoinRel.JoinType.JOIN_TYPE_INNER) {
+      shouldFallback = CHJoinValidateUtil.doValidate(condition)
+    }
+    if (shouldFallback) {
+      return false
+    }
+    super.doValidateInternal()
+  }
 }
 
 case class CHBroadcastHashJoinExecTransformer(
@@ -69,4 +83,21 @@ case class CHBroadcastHashJoinExecTransformer(
       newLeft: SparkPlan,
       newRight: SparkPlan): CHBroadcastHashJoinExecTransformer =
     copy(left = newLeft, right = newRight)
+
+  /*
+
+   */
+  override def doValidateInternal(): Boolean = {
+    var shouldFallback = false
+    if (substraitJoinType != JoinRel.JoinType.JOIN_TYPE_INNER) {
+      shouldFallback = CHJoinValidateUtil.doValidate(condition)
+    }
+    if (isNullAwareAntiJoin == true) {
+      shouldFallback = true
+    }
+    if (shouldFallback) {
+      return false
+    }
+    super.doValidateInternal()
+  }
 }