--- a/backends-velox/src/main/scala/io/glutenproject/backendsapi/velox/VeloxBackend.scala
+++ b/backends-velox/src/main/scala/io/glutenproject/backendsapi/velox/VeloxBackend.scala
@@ -166,6 +166,7 @@ object VeloxBackendSettings extends BackendSettings {
 
   override def recreateJoinExecOnFallback(): Boolean = true
   override def removeHashColumnFromColumnarShuffleExchangeExec(): Boolean = true
+  override def rescaleDecimalLiteral(): Boolean = true
 
   /**
    * Get the config prefix for each backend