--- a/backends-velox/src/main/scala/io/glutenproject/backendsapi/velox/VeloxInitializerApi.scala
+++ b/backends-velox/src/main/scala/io/glutenproject/backendsapi/velox/VeloxInitializerApi.scala
@@ -18,7 +18,7 @@ package io.glutenproject.backendsapi.velox
 
 import io.glutenproject.GlutenConfig
 import io.glutenproject.backendsapi.InitializerApi
-import io.glutenproject.utils.{VeloxDllLoaderUbuntu2004, VeloxDllLoaderUbuntu2204}
+import io.glutenproject.utils.{VeloxSharedlibraryLoaderUbuntu2004, VeloxSharedlibraryLoaderUbuntu2204, VeloxSharedlibraryLoaderCentos8, VeloxSharedlibraryLoaderCentos7, VeloxSharedlibraryLoader}
 import io.glutenproject.vectorized.{GlutenNativeExpressionEvaluator, JniLibLoader, JniWorkspace}
 import org.apache.commons.lang3.StringUtils
 import org.apache.spark.SparkConf
@@ -28,13 +28,16 @@ import scala.sys.process._
 class VeloxInitializerApi extends InitializerApi {
   def loadLibFromJar(load: JniLibLoader): Unit = {
       val system = "cat /etc/os-release".!!
-      if (system.contains("Ubuntu") && system.contains("20.04")) {
-        val loader = new VeloxDllLoaderUbuntu2004
-        loader.loadLib(load)
+      val loader = if (system.contains("Ubuntu") && system.contains("20.04")) {
+        new VeloxSharedlibraryLoaderUbuntu2004
       } else if (system.contains("Ubuntu") && system.contains("22.04")) {
-        val loader = new VeloxDllLoaderUbuntu2204
-        loader.loadLib(load)
+        new VeloxSharedlibraryLoaderUbuntu2204
+      } else if (system.contains("CentOS") && system.contains("8")){
+        new VeloxSharedlibraryLoaderCentos8
+      } else if (system.contains("CentOS") && system.contains("7")){
+        new VeloxSharedlibraryLoaderCentos7
       }
+      loader.asInstanceOf[VeloxSharedlibraryLoader].loadLib(load)
   }
 
   override def initialize(conf: SparkConf): Unit = {