--- a/backends-velox/src/main/scala/io/glutenproject/utils/VeloxExpressionUtil.scala
+++ b/backends-velox/src/main/scala/io/glutenproject/utils/VeloxExpressionUtil.scala
@@ -45,7 +45,9 @@ object VeloxExpressionUtil {
     TRANSLATE -> Set(EMPTY_TYPE),
     ADD_MONTHS -> Set(EMPTY_TYPE),
     DATE_FORMAT -> Set(EMPTY_TYPE),
-    TRUNC -> Set(EMPTY_TYPE)
+    TRUNC -> Set(EMPTY_TYPE),
+    SEQUENCE -> Set(EMPTY_TYPE),
+    POSEXPLODE -> Set(EMPTY_TYPE)
   )
 
   final val VELOX_AGGREGATE_FUNC_BLACKLIST: Map[String, Set[String]] = Map(