--- a/backends-velox/src/main/scala/io/glutenproject/utils/VeloxSharedlibraryLoader.scala
+++ b/backends-velox/src/main/scala/io/glutenproject/utils/VeloxSharedlibraryLoader.scala
@@ -2,7 +2,7 @@ package io.glutenproject.utils
 
 import io.glutenproject.vectorized.JniLibLoader
 
-trait VeloxDllLoader{
+trait VeloxSharedlibraryLoader{
     def loadLib(loader: JniLibLoader) : Unit = {}
 }