--- a/cpp/core/compute/Backend.h
+++ b/cpp/core/compute/Backend.h
@@ -22,8 +22,8 @@
 #include "memory/ArrowMemoryPool.h"
 #include "memory/ColumnarBatch.h"
 #include "operators/c2r/ArrowColumnarToRowConverter.h"
-#include "operators/shuffle/SplitterBase.h"
-#include "operators/shuffle/splitter.h"
+#include "shuffle/ArrowShuffleWriter.h"
+#include "shuffle/ShuffleWriter.h"
 #include "substrait/plan.pb.h"
 
 namespace gluten {
@@ -79,13 +79,14 @@ class Backend : public std::enable_shared_from_this<Backend> {
     return std::make_shared<ArrowColumnarToRowConverter>(rb, memory_pool);
   }
 
-  virtual std::shared_ptr<SplitterBase> makeSplitter(
+  virtual std::shared_ptr<ShuffleWriter> makeShuffleWriter(
       const std::string& partitioning_name,
       int num_partitions,
       const SplitOptions& options,
       const std::string& batchType) {
-    GLUTEN_ASSIGN_OR_THROW(auto splitter, Splitter::Make(partitioning_name, num_partitions, std::move(options)));
-    return splitter;
+    GLUTEN_ASSIGN_OR_THROW(
+        auto shuffle_writer, ArrowShuffleWriter::Make(partitioning_name, num_partitions, std::move(options)));
+    return shuffle_writer;
   }
 
   virtual std::shared_ptr<Metrics> GetMetrics(void* raw_iter, int64_t exportNanos) {