--- a/cpp/core/config/GlutenConfig.h
+++ b/cpp/core/config/GlutenConfig.h
@@ -27,6 +27,8 @@ const std::string kCaseSensitive = "spark.sql.caseSensitive";
 
 const std::string kSparkOffHeapMemory = "spark.gluten.memory.offHeap.size.in.bytes";
 
+const std::string kSparkTaskOffHeapMemory = "spark.gluten.memory.task.offHeap.size.in.bytes";
+
 const std::string kSparkBatchSize = "spark.sql.execution.arrow.maxRecordsPerBatch";
 
 std::unordered_map<std::string, std::string> getConfMap(JNIEnv* env, jbyteArray planArray);