… .github/workflows/velox_be.yml
… README.md
… backends-clickhouse/src/test/resources/tpcds-queries-output/q1.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q10.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q11.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q12.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q13.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q14a.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q14b.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q15.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q16.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q17.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q18.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q19.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q2.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q20.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q21.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q22.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q23a.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q24a.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q26.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q27.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q28.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q3.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q30.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q31.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q32.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q33.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q34.out
… backends-clickhouse/src/test/resources/tpcds-queries-output/q35.out