@FelixYBW 2023-03-23
> ```
> Optimize Velox to Internal Row. Previously we fill the destination column by column, which can hardly keep destination buffer data in L1 cache, especially when column number increases. But it needn't to flatten the row vector.
> 
> The new method flatten the row vector firstly, then similar as Arrow's C2R, we use a 3 loop to fill the buffer:
> 
> each batch is N rows
> for b in batches:
>   for c in b_column:
>     for r in rows:
>       buf[r][c]=b_column[c]
> 
> The expectation is both src and dst data is only fetched in L1 cache once
> ```
> 
> Perf: TPCH LITEITEM table 11 columns l_orderkey: int64 l_partkey: int64 l_suppkey: int64 l_linenumber: int32 l_quantity: double l_extendedprice: double l_discount: double l_tax: double l_shipinstruct: string l_shipmode: string l_comment: string
> 
> Arrow	before opt	after opt
> 1 thread	7.14s	11.3s	5.45s
> 32 threads	49.8s	105s	65.4s

TPCH