@JkSelf 2023-03-24
> > > Do you run at the jenkins? Can you backport decimal Velox to row from [#1176 (files)](https://github.com/oap-project/gluten/pull/1176/files#diff-467eb78d0235e3623ad443e886360015d85a71112637cdf5c1aa00721c100d41R174) to here?
> > 
> > 
> > @JkSelf Can you add UT for decimal VeloxColumnar2Row and VeloxSplitter?
> 
> @JkSelf Can you move your VeloxC2R to this PR?

Sure. I will move the VeloxColumnar2Row with decimal support here and add the UTs later.