@JkSelf 2023-03-24
@FelixYBW 
Filed [PR#1](https://github.com/FelixYBW/gluten/pull/1) to add short decimal and long decimal ut for VeloxColumnarToRow and VeloxSplitter. The VeloxColumnarToRow can pass the ut. But it seems there has some issue in VeloxSplitter with decimal type.